Struct anoma_shared::ledger::gas::BlockGasMeter

pub struct BlockGasMeter { /* fields omitted */ }

Gas metering in a block. Tracks the gas in a current block and a current transaction.

Implementations

impl BlockGasMeter

pub fn add(&mut self, gas: u64) -> Result<()>

Add gas cost for the current transaction. It will return error when the consumed gas exceeds the transaction gas limit, but the state will still be updated.

pub fn add_base_transaction_fee(&mut self, bytes_len: usize) -> Result<()>

Add the base transaction fee and the fee per transaction byte that’s charged the moment we try to apply the transaction.

pub fn add_compiling_fee(&mut self, bytes_len: usize) -> Result<()>

Add the compiling cost proportionate to the code length

pub fn finalize_transaction(&mut self) -> Result<u64>

Add the transaction gas to the block’s total gas. Returns the transaction’s gas cost and resets the transaction meter. It will return error when the consumed gas exceeds the block gas limit, but the state will still be updated.

pub fn reset(&mut self)

Reset the gas meter.

pub fn get_current_transaction_gas(&self) -> u64

Get the total gas used in the current transaction.

pub fn add_vps_gas(&mut self, vps_gas: &VpsGas) -> Result<()>

Add the gas cost used in validity predicates to the current transaction.