Trait anoma_shared::ledger::native_vp::NativeVp

pub trait NativeVp {
    const ADDR: InternalAddress;

    fn init_genesis_storage<DB, H>(storage: &mut Storage<DB, H>)
    where
        DB: DB + for<'iter> DBIter<'iter>,
        H: StorageHasher;

    fn validate_tx<DB, H>(
        ctx: &mut Ctx<'_, DB, H>, 
        tx_data: &[u8], 
        keys_changed: &HashSet<Key>, 
        verifiers: &HashSet<Address>
    ) -> Result<bool>
    where
        DB: DB + for<'iter> DBIter<'iter>,
        H: StorageHasher;
}

A native VP module should implement this module and add its initialization call to init_genesis_storage.

Associated Constants

const ADDR: InternalAddress

The address of this VP

Loading content...
Required methods

fn init_genesis_storage<DB, H>(storage: &mut Storage<DB, H>) where
DB: DB + for<'iter> DBIter<'iter>,
H: StorageHasher,

Initialize storage in the genesis block

fn validate_tx<DB, H>(
ctx: &mut Ctx<'_, DB, H>,
tx_data: &[u8],
keys_changed: &HashSet<Key>,
verifiers: &HashSet<Address>
) -> Result<bool> where
DB: DB + for<'iter> DBIter<'iter>,
H: StorageHasher,

Run the validity predicate. This function can call methods on the Ctx argument to interact with the host structures.

Loading content...