Module anoma_shared::ledger::storage

Ledger’s state storage with key-value backed store and a merkle tree

Modules

mockdb
DB mock for testing

testing
Helpers for testing components that depend on storage

types
The key and values that may be persisted in a DB.

write_log
Write log is temporary storage for modifications performed by a transaction. before they are committed to the ledger’s storage.

Structs

BlockState
The block’s state as stored in the database.

BlockStorage
The block storage data

MerkleRoot
The root hash of the merkle tree as bytes

Storage
The storage data

Enums

Error

Traits

DB
A database backend.

DBIter
A database prefix iterator.

StorageHasher
The storage hasher used for the merkle tree.

Type Definitions

Result
A result of a function that may fail