Module anoma_shared::types

The key and values that may be persisted in a DB.

Re-exports

pub use address::Address;

pub use address::EstablishedAddress;

pub use address::ImplicitAddress;

Modules

address
Implements transparent addresses as described in Accounts Addresses.

intent

internal
Shared internal types between the host env and guest (wasm).

key

token
A basic fungible token

Structs

BlockHash

BlockHeight

Key

UpdateVp
A tx data type to update an account’s validity predicate

Enums

DbKeySeg

Error

Constants

BLOCK_HASH_LENGTH

CHAIN_ID_LENGTH

KEY_SEGMENT_SEPARATOR

RESERVED_ADDRESS_PREFIX

RESERVED_VP_KEY

VP_KEY_PREFIX

Traits

KeySeg
Represents a segment in a path that may be used as a database key

Type Definitions

Result