Function anoma_shared::vm::host_env::tx_iter_next

pub fn tx_iter_next<MEM, DB, H>(
    env: &TxEnv<'_, MEM, DB, H>, 
    iter_id: u64
) -> Result<i64, TxRuntimeError> where
    MEM: VmMemory,
    DB: DB + for<'iter> DBIter<'iter>,
    H: StorageHasher, 

Storage prefix iterator next function exposed to the wasm VM Tx environment. It will try to read from the write log first and if no entry found then from the storage.

Returns -1 when the key is not present, or the length of the data when the key is present (the length may be 0).