Function anoma_shared::vm::host_env::vp_result_buffer

pub fn vp_result_buffer<MEM, DB, H, EVAL>(
    env: &VpEnv<'_, MEM, DB, H, EVAL>, 
    result_ptr: u64
) where
    MEM: VmMemory,
    DB: DB + for<'iter> DBIter<'iter>,
    H: StorageHasher,
    EVAL: VpEvalRunner, 

This function is a helper to handle the first step of reading var-len values from the host.

In cases where we’re reading a value from the host in the guest and we don’t know the byte size up-front, we have to read it in 2-steps. The first step reads the value into a result buffer and returns the size (if any) back to the guest, the second step reads the value from cache into a pre-allocated buffer with the obtained size.