Trait anoma_shared::vm::host_env::VpEvaluator

pub trait VpEvaluator {
    type Db: DB + for<'iter> DBIter<'iter>;
    type H: StorageHasher;
    type Eval: VpEvaluator;
    fn eval(
        &self, 
        ctx: VpCtx<'static, Self::Db, Self::H, Self::Eval>, 
        vp_code: Vec<u8>, 
        input_data: Vec<u8>
    ) -> HostEnvResult;
}

A Validity predicate runner for calls from the vp_eval function.

Associated Types

type Db: DB + for<'iter> DBIter<'iter>

Storage DB type

type H: StorageHasher

Storage hasher type

type Eval: VpEvaluator

Recursive VP evaluator type

Loading content...
Required methods

fn eval(
&self,
ctx: VpCtx<'static, Self::Db, Self::H, Self::Eval>,
vp_code: Vec<u8>,
input_data: Vec<u8>
) -> HostEnvResult

Evaluate a given validity predicate code with the given input data. Currently, we can only evaluate VPs using WASM runner with WASM memory.

Invariant: Calling VpEvalRunner::eval from the VP is synchronous as it shares mutable access to the host context with the VP.

Loading content...