Trait anoma_shared::vm::memory::VmMemory

pub trait VmMemory: Clone + Send + Sync {
    type Error: Error + Sync + Send + 'static;
    fn read_bytes(
        &self, 
        offset: u64, 
        len: usize
    ) -> Result<(Vec<u8>, u64), Self::Error>;

    fn write_bytes(
        &self, 
        offset: u64, 
        bytes: impl AsRef<[u8]>
    ) -> Result<u64, Self::Error>;

    fn read_string(
        &self, 
        offset: u64, 
        len: usize
    ) -> Result<(String, u64), Self::Error>;

    fn write_string(
        &self, 
        offset: u64, 
        string: String
    ) -> Result<u64, Self::Error>;
}

Abstract representation of virtual machine’s memory.

Associated Types

type Error: Error + Sync + Send + 'static

Error type for the methods’ results.

Loading content...
Required methods

fn read_bytes(
&self,
offset: u64,
len: usize
) -> Result<(Vec<u8>, u64), Self::Error>

Returns bytes read from memory together with the associated gas cost.

fn write_bytes(
&self,
offset: u64,
bytes: impl AsRef<[u8]>
) -> Result<u64, Self::Error>

Write bytes to memory. Returns the gas cost.

fn read_string(
&self,
offset: u64,
len: usize
) -> Result<(String, u64), Self::Error>

Returns string read from memory together with the associated gas cost.

fn write_string(&self, offset: u64, string: String) -> Result<u64, Self::Error>

Write string to memory. Returns the gas cost.

Loading content...