Module anoma_shared::vm::types

VM types can be passed between the host and guest via wasm linear memory.

These are either:

Module call types The module call inputs are passed host-to-guest.

Execution environment types The environment inputs are passed guest-to-host and outputs back from host-to-guest.

Structs

KeyVal
Key-value pair represents data from account’s subspace

VpInput
Input for validity predicate wasm module call

Type Definitions

MatchmakerInput
Input for matchmaker wasm module call

TxInput
Input for transaction wasm module call