Module anoma_shared::vm::wasm::memory

Wasm memory is used for bi-directionally passing data between the host and a wasm instance.

Structs

FilterCallInput
Input data for matchmaker filter wasm call

Limit
A custom [Tunables] to set a WASM memory limits.

MatchmakerCallInput
Input data for matchmaker wasm call

TxCallInput
Input data for transaction wasm call

VpCallInput
Input data for validity predicate wasm call

WasmMemory
The wasm memory

Enums

Error

Constants

FILTER_MEMORY_INIT_PAGES
Initial pages in matchmaker filter memory

MATCHMAKER_MEMORY_INIT_PAGES
Initial pages in matchmaker memory

TX_MEMORY_INIT_PAGES
Initial pages in tx memory

TX_MEMORY_MAX_PAGES
Mamixmum pages in tx memory

VP_MEMORY_INIT_PAGES
Initial pages in VP memory

VP_MEMORY_MAX_PAGES
Mamixmum pages in VP memory

Functions

prepare_filter_memory
Prepare memory for instantiating a filter module

prepare_matchmaker_memory
Prepare memory for instantiating a matchmaker module

prepare_tx_memory
Prepare memory for instantiating a transaction module

prepare_vp_memory
Prepare memory for instantiating a validity predicate module

tx_limit
A Limit with memory limit setup for transaction WASM execution.

vp_limit
A Limit with memory limit setup for validity predicate WASM execution.

write_filter_inputs
Write matchmaker filter inputs into wasm memory

write_matchmaker_inputs
Write matchmaker inputs into wasm memory

write_tx_inputs
Write transaction inputs into wasm memory

write_vp_inputs
Write validity predicate inputs into wasm memory

Type Definitions

Result
Result of a function that may fail