Module anoma_shared::vm::wasm::run

Wasm runners

Structs

VpEvalWasm
Validity predicate wasm evaluator for eval host function calls.

Enums

Error

Functions

matchmaker
Execute a matchmaker code.

matchmaker_filter
Execute a matchmaker filter code to check if it accepts the given intent.

tx
Execute a transaction code. Returns the set verifiers addresses requested by the transaction.

vp
Execute a validity predicate code. Returns whether the validity predicate accepted storage modifications performed by the transaction that triggered the execution.

Type Definitions

Result
Result for functions that may fail